format eslint

+++ apps/transactions-service/src/frequencies/frequencies.service.ts
@@ -1,29 +1,35 @@
-import { Injectable } from '@nestjs/common';
-import { CreateFrequencyDto } from './dto/create-frequency.dto';
-import { UpdateFrequencyDto } from './dto/update-frequency.dto';
-import { FrequenciesRepository } from './repository/frequencies.repository';
+import { Injectable } from "@nestjs/common";
+import { Frequency } from "src/generated/prisma/client";
+
+import { CreateFrequencyDto } from "./dto/create-frequency.dto";
+import { UpdateFrequencyDto } from "./dto/update-frequency.dto";
+import { FrequenciesRepository } from "./repository/frequencies.repository";
 
 @Injectable()
 export class FrequenciesService {
-  constructor(private readonly repository: FrequenciesRepository) { }
-  
-  create(createFrequencyDto: CreateFrequencyDto) {
+  constructor(private readonly repository: FrequenciesRepository) {}
+
+  create(createFrequencyDto: CreateFrequencyDto): Promise<Frequency> {
     return this.repository.create(createFrequencyDto);
   }
 
-  findAll() {
+  findAll(): Promise<Frequency[]> {
     return this.repository.getAll();
   }
 
-  findOne(id: string) {
-    return this.repository.findOne(id);
+  findOne(id: string): Promise<Frequency | null> {
+    return this.repository.getById(id);
   }
 
-  update(id: string, updateFrequencyDto: UpdateFrequencyDto) {
-    return this.repository.update(id, updateFrequencyDto);
+  update(id: string, updateFrequencyDto: UpdateFrequencyDto): Promise<Frequency> {
+    return this.repository.updateOne(id, updateFrequencyDto);
   }
 
-  remove(id: string) {
+  async remove(id: string): Promise<Frequency | null> {
+    const frequency = await this.repository.getById(id);
+
+    if (!frequency) return null;
+
     return this.repository.remove(id);
   }
 }
